requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
firebase-admin==6.2.0
pyrebase4==4.7.1
google-generativeai==0.3.1
gunicorn==21.2.0
python-dotenv==1.0.0
requests<2.30,>=2.19.1
beautifulsoup4==4.12.2
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7